pyproject.toml
[project]
name = "clangen-catmaker-main"
version = "1.0.0"
description = ""
authors = [{ name = "Chinch-Bug", email = "noemail@null.com" }]
readme = "README.md"

keywords = []
requires-python = ">=3.10"

dependencies = [
    "pygame-ce~=2.2",
    "pygame-gui~=0.6",
    "ujson~=5.7",
    "bidict~=0.23",
    "tomli>=2.4.1",
]

[dependency-groups]
build = [
    "pillow~=9.5",
    "pyinstaller~=6.6",
    "setuptools~=67.7",
    "wheel~=0.40",
]
